Revert: save-compressed-image feature.
This feature needs to be better thought out.  How is the user expected to restore these saved images?  Also, creating a new top level directory for every feature is a no-go, it polutes the directory tree making it harder for the user to orient themselves in the sourcecode.

What exactly is the goal of this feature?  Is the user then going to copy the entire subuser directory over to another machine and type some command like "subuser install-from-saved-images"?  I don't understand the user interface that is expected here...

Tim