Allow running subuser as root user
This allows you to not have your normal user be a member of the docker group #131 #223 . This is extremrely messy, non-documented, non-tested code.

In order to use, create a `/root/.subuser` directory and put a `/root/.subuser/config.json` file inside:

{
 "user":"timothy"
,"subuser-home-dirs-dir":"/home/timothy/subuser-homes"
}

Replace `timothy` with your username.

This will alow you to run subuser as root, but have your subuser's run as the user timothy.